Don't assume the main application is mounted at /.
No, you cannot use the config values for anything other than setting routes.

If you want to link to something, you should ask Flask for url_for("something"),
not blindly assume the path in the config is going to work. Because if the main
application is mounted at, say, '/flask', it won't.